refactor(amplify-provider-awscloudformation): renamed the files to fit
