+++ packages/amplify-provider-awscloudformation/src/resource-package/constants.ts
