+++ packages/amplify-provider-awscloudformation/src/resource-package/types.ts
